requirements.txt
langchain
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
tiktoken==0.8.0
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
openai==1.40.0
langchain_openai==0.2.1
langchain_community==0.3.0
pymupdf